suppress warning
git-svn-id: https://svn.salilab.org/imp/trunk@4469 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ kernel/src/ConfigurationSet.cpp
@@ -71,7 +71,8 @@ void ConfigurationSet::set_base() {
 }
 
 void ConfigurationSet::set_structure(int i) {
-  IMP_USAGE_CHECK(i < get_number_of_configurations()
+  IMP_USAGE_CHECK(static_cast<unsigned int>(i)
+                  < get_number_of_configurations()
                   && i >= -1,
                   "Invalid configuration requested.", IndexException);
   set_base();
